qgsframegraph: move framegraph files and dependencies to dedicated directory
