+++ src/3d/framegraph/qgsambientocclusionblurentity.h
