+++ src/3d/framegraph/qgsambientocclusionrenderentity.cpp
