+++ src/3d/framegraph/qgsambientocclusionrenderentity.h
